Refactors Thumbnail into component
